- Move mixin to it's own file
