+++ src/sparseml/core/modifier/modifier.py
@@ -12,106 +12,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
-from typing import Any, Generator, Optional, Tuple
+from typing import Optional
 
 from pydantic import BaseModel
 
 from sparseml.core.event import Event, EventType
 from sparseml.core.framework_object import MultiFrameworkObject
-from sparseml.core.logger import LoggerManager
 from sparseml.core.modifier.base import ModifierInterface
+from sparseml.core.modifier.mixins import ModelLoggingMixin
 from sparseml.core.state import State
 
 
-__all__ = ["Modifier"]
-
-
-class ModelLoggingMixin_:
-    """
-    A mixin that adds model level logging functionality
-    """
-
-    def log_model_info(self, state: State, event: Event):
-        """
-        Log model level info to the logger
-        Relies on `state.model` having a `loggable_items` method
-        that returns a generator of tuples of the loggable item
-        name and value. Only logs on BATCH_END type events at the
-        end of an epoch. Also relies on `state.loggers` being a
-        `LoggerManager` instance.
-
-        :param state: The current state of sparsification
-        :param event: The event to update the modifier with
-        """
-
-        if not self._should_log_model_info(state, event):
-            return
-        self._log_epoch(logger_manager=state.loggers, epoch=int(event.current_index))
-        self._log_model_loggable_items(
-            logger_manager=state.loggers,
-            loggable_items=state.model.loggable_items(),
-            epoch=event.current_index,
-        )
-
-    def _should_log_model_info(self, state: State, event: Event) -> bool:
-        """
-        Check if we should log model level info
-        Criteria:
-            - model has a loggable_items method
-            - event is of type BATCH_END
-            - event is at the end of an epoch
-            - state has a logger manager
-
-
-        :param state: The current state of sparsification
-        :param event: The event to update the modifier with
-        :return: True if we should log model level info, False otherwise
-        """
-        return (
-            hasattr(state.model, "loggable_items")
-            and event.type_ == EventType.BATCH_END
-            and isinstance(state.loggers, LoggerManager)
-            and state.loggers.epoch_to_step(
-                epoch=event.current_index, steps_per_epoch=event.steps_per_epoch
-            )
-            % event.steps_per_epoch
-            == 0
-        )
-
-    def _log_epoch(self, logger_manager: LoggerManager, epoch: int):
-        """
-        Log the epoch to the logger_manager
-
-        :param logger_manager: The logger manager to log to
-        :param epoch: The epoch to log
-        """
-        epoch_str = f"Epoch: #{epoch}"
-        logger_manager.log_string(tag="Epoch", string=f"{epoch_str:=^20}", step=epoch)
-
-    def _log_model_loggable_items(
-        self,
-        logger_manager: LoggerManager,
-        loggable_items: Generator[Tuple[str, Any], None, None],
-        epoch: float,
-    ):
-        """
-        Log the model level loggable items to the logger_manager
-
-        :param logger_manager: The logger manager to log to
-        :param loggable_items: The loggable items to log, must be a generator of tuples
-            of the loggable item name and value
-        :param epoch: The epoch to log
-        """
-        for loggable_item in loggable_items:
-            log_tag, log_value = loggable_item
-            if isinstance(log_value, dict):
-                logger_manager.log_scalars(tag=log_tag, values=log_value, step=epoch)
-            else:
-                logger_manager.log_string(tag=log_tag, string=log_value, step=epoch)
-
-
-class Modifier(BaseModel, ModifierInterface, MultiFrameworkObject, ModelLoggingMixin_):
+class Modifier(BaseModel, ModifierInterface, MultiFrameworkObject, ModelLoggingMixin):
     """
     A base class for all modifiers to inherit from.
     Modifiers are used to modify the training process for a model.
